+++ squirrel/driver/__init__.py
@@ -1,6 +1,6 @@
 from squirrel.driver.csv import CsvDriver
-from squirrel.driver.data_frame_file import DataFrameFileDriver
-from squirrel.driver.driver import DataFrameDriver, Driver, IterDriver, MapDriver
+from squirrel.driver.data_frame import DataFrameDriver
+from squirrel.driver.driver import Driver, IterDriver, MapDriver
 from squirrel.driver.file import FileDriver
 from squirrel.driver.feather import FeatherDriver
 from squirrel.driver.json import JsonDriver
@@ -15,7 +15,6 @@ from squirrel.driver.zarr import ZarrDriver
 __all__ = [
     "CsvDriver",
     "DataFrameDriver",
-    "DataFrameFileDriver",
     "Driver",
     "FeatherDriver",
     "FileDriver",
